Reworked version of randomwalk. Features:
- added requirements.txt and setup.py for easy_install;
- reworked clui interface;

Current version doesn't contain any slurm sbatch wrappers.
