+++ README.md
@@ -1,6 +1,10 @@
- randomwalk.py
- input parameters are: number of MC runs, index in the sp500./csv file, Quandl API key
- the file s&p500.csv should be in the same directory
- the output file is produced in the directory /tmp/outdata in the format <stock symbol>.csv
- each output file includes a random position (number of shares) between 100 and 10000
+# vardemo1
 
+This package provides `randomwalk` tool which does Monte-Carlo simulation
+of stock prices behaviour based on data from
+[quandl.com](https://www.quandl.com).
+
+Stock company labels could be found in `companies.csv` file.
+
+### Notes
+  - You will need `Tkinter` installed to use `--plot` option of the tool.
